refactor(core): simplify order setup and relocate selection activity helper
Keep selection-state logic with the selection utilities and remove redundant order validation.
